Moved wave to own component & Changed Category Context placement
